Added the final project files.
I finished the training of the neural network, and saved it as bot.hdf5, added tags list and words list which are binary files that areserialisation python list objects. I made the script retrain and save the model and pickled files if it canot load them. Con.json is a json files that contain sample intents and phrases that represent them, and what are the appropriate response.